Add Preivew with submit and alert working, still working on select item
