+++ src/components/section/preivew/index.js
@@ -1,7 +1,15 @@
 import SectionPreview from './SectionPreview'
 import Constant from './Constant'
+import BuildField from './BuildField'
+import BuildItem from './BuildItem'
+import BuildList from './BuildList'
+import BuildSubmitBtn from './BuildSubmitBtn'
 
 export {
+    BuildField,
+    BuildItem,
+    BuildList,
+    BuildSubmitBtn,
     Constant,
     SectionPreview
 }
